Add crud services

+++ src/main/java/ru/skb/cos/repository/TemplateRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import ru.skb.cos.model.entity.TemplateEntity;
 
 @RepositoryRestResource(path = "Templates")
-public interface TemplateRepository extends CrudRepository<TemplateEntity, Integer> {
+public interface TemplateRepository extends CrudRepository<TemplateEntity, Long> {
 }
